Extract fields framework
